requirements.txt
pillow 
numpy
pandas